tools/pdl2palaestrai/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdl2palaestrai"
version = "0.1.0"
description = "Convert PROVIDER PDL YAML scenarios into palaestrai experiment inputs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
]
authors = [
  {name = "PROVIDER"}
]

[project.scripts]
pdl2palaestrai = "pdl2palaestrai.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
